# pos-carts/target/generated-sources/openapi/src/main/java/com/micropos/carts/api/CartApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (5.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.micropos.carts.api;

import com.micropos.carts.dto.CartDto;
import com.micropos.carts.dto.ErrorDto;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2022-05-23T06:27:44.297465-07:00[America/Los_Angeles]")
@Validated
@Tag(name = "cart", description = "the cart API")
public interface CartApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /cart/add/{cartId}/{productId} : Add a product to cart
     *
     * @param cartId Cart Id (required)
     * @param productId Product ID (required)
     * @return Success (status code 200)
     *         or Cart not found (status code 404)
     */
    @Operation(
        operationId = "addProduct",
        summary = "Add a product to cart",
        tags = { "cart" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Success", content = @Content(mediaType = "application/json", schema = @Schema(implementation =  CartDto.class))),
            @ApiResponse(responseCode = "404", description = "Cart not found", content = @Content(mediaType = "application/json", schema = @Schema(implementation =  ErrorDto.class)))
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/cart/add/{cartId}/{productId}",
        produces = { "application/json" }
    )
    default ResponseEntity<CartDto> addProduct(
        @Parameter(name = "cartId", description = "Cart Id", required = true, schema = @Schema(description = "")) @PathVariable("cartId") Long cartId,
        @Parameter(name = "productId", description = "Product ID", required = true, schema = @Schema(description = "")) @PathVariable("productId") Long productId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"cartId\" : 0, \"items\" : [ { \"quantity\" : 1.4658129805029452, \"productId\" : 6 }, { \"quantity\" : 1.4658129805029452, \"productId\" : 6 } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /cart/checkout/{cartId} : checkout a cart
     *
     * @param cartId The id of the cart to retrieve (required)
     * @return Success (status code 200)
     *         or Cart not found (status code 404)
     */
    @Operation(
        operationId = "checkout",
        summary = "checkout a cart",
        responses = {
            @ApiResponse(responseCode = "200", description = "Success"),
            @ApiResponse(responseCode = "404", description = "Cart not found", content = @Content(mediaType = "application/json", schema = @Schema(implementation =  ErrorDto.class)))
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/cart/checkout/{cartId}",
        produces = { "application/json" }
    )
    default ResponseEntity<Void> checkout(
        @Parameter(name = "cartId", description = "The id of the cart to retrieve", required = true, schema = @Schema(description = "")) @PathVariable("cartId") Long cartId
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /cart/create : create a cart
     *
     * @return Success (status code 200)
     *         or Cart not found (status code 404)
     */
    @Operation(
        operationId = "createCart",
        summary = "create a cart",
        tags = { "cart" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Success", content = @Content(mediaType = "application/json", schema = @Schema(implementation =  CartDto.class))),
            @ApiResponse(responseCode = "404", description = "Cart not found", content = @Content(mediaType = "application/json", schema = @Schema(implementation =  ErrorDto.class)))
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/cart/create",
        produces = { "application/json" }
    )
    default ResponseEntity<CartDto> createCart(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"cartId\" : 0, \"items\" : [ { \"quantity\" : 1.4658129805029452, \"productId\" : 6 }, { \"quantity\" : 1.4658129805029452, \"productId\" : 6 } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}